Added a secondary structure renderer in 2D
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/protein_cuda/trunk@1481 8215d0b1-3b2f-4a13-8d91-4b3e73067666

+++ src/CUDAMarchingCubes.h
@@ -141,7 +141,7 @@ public:
      *
      * @return The voxel count
      */
-    unsigned int GetVoxelCount() { return numVoxels; }
+    unsigned int GetVoxelCount() { return static_cast<unsigned int>(numVoxels); }
 
     /**
      * Returns wether a color table was computed for the surface or not.
